# tyst-api-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tyst_api_rest",
    "description": "TYST REST API",
    "license": {
      "name": "Apache-2.0 WITH FWM-Exception-1.0.0",
      "identifier": "Apache-2.0 WITH FWM-Exception-1.0.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/kem/{algorithm}/decapsulate": {
      "post": {
        "tags": [
          "kem_resources"
        ],
        "summary": "Derive a shared secret though KEM decapsulation.",
        "description": "Derive the same shared secret as the other party by decapsulating the\nrecieved cipher text with the decapsulation key.",
        "operationId": "kem_decapsulate",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "decapsulation_key",
                  "cipher_text_b64"
                ],
                "properties": {
                  "cipher_text_b64": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32",
                      "minimum": 0
                    }
                  },
                  "decapsulation_key": {
                    "type": "object",
                    "required": [
                      "key_material_b64"
                    ],
                    "properties": {
                      "confinement_id": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Identifier used by provider to locate the correct confinement."
                      },
                      "confinement_type": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Type of confinement. E.g. \"soft\"."
                      },
                      "key_material_b64": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "format": "int32",
                          "minimum": 0
                        },
                        "description": "Base64 encoded key material"
                      },
                      "object_reference": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "description": "Identifier used by provider to locate the correct object."
                      }
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "shared_secret_b64"
                  ],
                  "properties": {
                    "shared_secret_b64": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failure"
          }
        }
      }
    },
    "/kem/{algorithm}/encapsulate": {
      "post": {
        "tags": [
          "kem_resources"
        ],
        "summary": "Generate a shared secret and encapsulate it.",
        "description": "Generate a shared secret and encapsulate the shared secret using the KEM\nencapsulation key. The encapsulated shared secret is returned both in plain\nand as cipher_text. The cipher_text can be sent to the other party.",
        "operationId": "kem_encapsulate",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "encapsulation_key_b64"
                ],
                "properties": {
                  "encapsulation_key_b64": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "cipher_text_b64",
                    "shared_secret_b64"
                  ],
                  "properties": {
                    "cipher_text_b64": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                      }
                    },
                    "shared_secret_b64": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failure"
          }
        }
      }
    },
    "/kem/{algorithm}/keygen": {
      "post": {
        "tags": [
          "kem_resources"
        ],
        "summary": "Start generation of KEM en-/decapsulation keys.",
        "description": "Start generation of a pair of suitable encapsulation and decapsulation keys\nfor the Key Encapsulation Mechanism (KEM) algorithm.",
        "operationId": "kem_keygen",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "description": "The KEM algorithm.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Key generation has started. Redirecting to result resource."
          },
          "404": {
            "description": "Failed to start"
          },
          "429": {
            "description": "Too many requests are already in progress"
          }
        }
      }
    },
    "/kem/{algorithm}/keygen/{generation_request_id}": {
      "get": {
        "tags": [
          "kem_resources"
        ],
        "summary": "Get generated KEM en-/decapsulation keys.",
        "description": "Get generated pair of suitable encapsulation and decapsulation keys for the\nKey Encapsulation Mechanism (KEM) algorithm.",
        "operationId": "kem_keygen_result",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "description": "The KEM algorithm.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "generation_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "decapsulation_key",
                    "encapsulation_key_b64"
                  ],
                  "properties": {
                    "decapsulation_key": {
                      "type": "object",
                      "required": [
                        "key_material_b64"
                      ],
                      "properties": {
                        "confinement_id": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Identifier used by provider to locate the correct confinement."
                        },
                        "confinement_type": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Type of confinement. E.g. \"soft\"."
                        },
                        "key_material_b64": {
                          "type": "array",
                          "items": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 0
                          },
                          "description": "Base64 encoded key material"
                        },
                        "object_reference": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "description": "Identifier used by provider to locate the correct object."
                        }
                      }
                    },
                    "encapsulation_key_b64": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "307": {
            "description": "Not ready yet"
          },
          "404": {
            "description": "Key pair generation failed or was never started."
          }
        }
      }
    },
...
  },
...
}